% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_heritage.R
\name{.clade_probabilities}
\alias{.clade_probabilities}
\title{Internal function: Calculate probabilities of matching states within a clade}
\usage{
.clade_probabilities(state)
}
\arguments{
\item{state}{the state to condition on}

\item{clade_states}{A named vector of trait values}
}
\value{
numerator: the number of matching pairs within a clade

denominator: The number of taxa pairs within a clade
}
\description{
Internal function: Calculate probabilities of matching states within a clade
}
