% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_heritage.R
\name{.get_clades}
\alias{.get_clades}
\title{Internal function: Cuts a phylogeny at a given point}
\usage{
.get_clades(tree, cut = NULL, k = NULL, desc)
}
\arguments{
\item{tree}{is a dated phylogenetic tree with branch lengths stored as a phylo object (as in the ape package).}

\item{cut}{the slice time}

\item{k}{number of slices}
}
\value{
a data frame showing the clades for each taxa
}
\description{
This function has been adapted from phyloregion, but included separately here to avoid unnecessary dependencies from that package
Please cite Daru, B. H., Karunarathne P., & Schliep K. (2020), phyloregion: R package for biogeographic regionalization and macroecology. Methods in Ecology and Evolution, 11: 1483-1491. doi:10.1111/2041-210X.13478
if using this function.
}
