% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_heritage.R
\name{.slice_tree}
\alias{.slice_tree}
\title{Internal Function: Identify clades at all generations}
\usage{
.slice_tree(tree, n_generations)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{n_generations}{A number showing the number of generations to calculate for the tree. User should calculate this based on their branch lengths}
}
\value{
clades: a data.frame showing the clades for each taxa within each generation.
}
\description{
Internal Function: Identify clades at all generations
}
