% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permute_trait_heritage}
\alias{permute_trait_heritage}
\title{Calculate the true probabilities of shared traits and the permuted probability}
\usage{
permute_trait_heritage(
  tree,
  trait,
  generation_time,
  n_permutations = 100,
  seed = 937
)
}
\arguments{
\item{tree}{a single phylogeny}

\item{trait}{a vector of traits, named for each taxa}

\item{generation_time}{the time between generations to cut the phylogeny at}

\item{n_permutations}{the number of permutations to run as a comparison (default = 100)}
}
\value{
a list containing the estimated probability, and the permuted probabilities.
}
\description{
Calculate the true probabilities of shared traits and the permuted probability
}
