% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trmca.R
\name{tmrca}
\alias{tmrca}
\title{Calculate the Time since Most Recent Common Ancestor across a tree posterior}
\usage{
tmrca(trees, states, progress = TRUE)
}
\arguments{
\item{trees}{a phylogeny or posterior of phylogeny of class phylo or multiphylo}

\item{states}{a named vector of states to calculate clades from. Names must match tips.}
}
\value{
a data.frame containing the dates of the MRCA of state clades for each tree in the posterior
}
\description{
Calculate the Time since Most Recent Common Ancestor across a tree posterior
}
