% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_evolution.R
\name{trait_coevolution}
\alias{trait_coevolution}
\title{Calculate the relative probability of a shared trait and distance}
\usage{
trait_coevolution(tree, trait, distance_matrix, generation_time, cut_off)
}
\arguments{
\item{tree}{a phylogenetic tree of classe phylo}

\item{trait}{a numeric or character trait for each taxa}

\item{distance_matrix}{a n by n matrix of distances between all taxa}

\item{generation_time}{the number of generations to calculate the probaiblity for across the phylogeny}

\item{cut_off}{The distance cut off value (used to discretize distance)}
}
\value{
a list. time_df shows the probability for each paired category over time. pairs_df shows the data for each viable pair in the dataset.
}
\description{
Calculate the relative probability of a shared trait and distance
}
