% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_heritage.R
\name{trait_heritage}
\alias{trait_heritage}
\title{Calculate the heritage of a trait along a single phylogeny}
\usage{
trait_heritage(tree, trait, generation_time)
}
\arguments{
\item{tree}{a single phylogeny}

\item{trait}{a vector of traits, named for each taxa}

\item{generation_time}{the number of times to cut a tree and calculate the probability of trait heritage}
}
\value{
a dataframe containing the probability of shared languages within each generation
}
\description{
Calculate the heritage of a trait along a single phylogeny
}
