% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_heritage.R
\name{trait_heritage_specific}
\alias{trait_heritage_specific}
\title{Calculate the heritage of a specific character trait along a single phylogeny}
\usage{
trait_heritage_specific(tree, trait, generation_time, condition = 1)
}
\arguments{
\item{tree}{a single phylogeny}

\item{trait}{a vector of traits, named for each taxa}

\item{generation_time}{how frequently the trait probability should be calculated. Must be less than the height of the phylogeny.}
}
\value{
a list containing two dataframes. by_trait contains the probability calculation at each timestep for each level of the trait. summary shows the probability across trait values for each time step.
}
\description{
Calculate the heritage of a specific character trait along a single phylogeny
}
