% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_between.R
\name{within_between}
\alias{within_between}
\title{Calculate within and between variance for a trait in a tree}
\usage{
within_between(tree, generation_time, trait)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{generation_time}{the generation time to calcualte across the phylogeny}

\item{trait}{the trait of interest (numeric or character)}
}
\description{
Calculate within and between variance for a trait in a tree
}
